Add README.md content and javascript getComputerChoice function
adds description for this current project and tools use in this project
in README.md file.

adds function of getComputerChoice and inside that function declare
variable of randomnumber to generate a random number between 0 to 2,
variables of choices, computerChoice, and return computerChoice to map
the random number to 'rock', 'paper', or 'scissors'.

add console.log to test the function by logging the result to the
console.