Add function game to o play a best-of-five game that keeps score and reports a winner or loser at the end.
add function playIndividual Round to play a round game.
add const playerSelection to input prompt for palyer to play game.
add const result to display output game in one round and add console.log
to display output per round.

make a logic for Handling ties, wins, and loses in the game to go to
the next round.

make a logic for continue palying until 5 non-tie.

make a function to declare the final winner. call function of
playIndividualRound to start the first round. call function of game() to
start playing.